+++ app/src/contants.ts
@@ -1,9 +1,7 @@
 export const TIDAL_ITEMS_PER_PAGE = 18;
-export const TIDAL_API_URL = "https://api.tidal.com";
 export const TIDAL_ALBUM_URL = "http://www.tidal.com/album/";
 
 export const LOCALSTORAGE_TOKEN_KEY = "tidarr-token";
 export const LOCALSTORAGE_REDIRECT_URL = "tidarr-redirect";
 export const LOCALSTORAGE_UPDATE_WARNING = "tidarr-update-warn";
-
 export const LOCALSTORAGE_DISPLAY_MODE = "tidarr-display-mode";